keyword入力についてモジュール分割しました
