--- app/src/main/java/com/example/gourmetsearchercompose/ui/screen/inputkeyword/component/range/RangeList.kt
@@ -1,38 +1,0 @@
-package com.example.gourmetsearchercompose.ui.screen.inputkeyword.component.range
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.example.gourmetsearchercompose.R
-
-/**
- * 範囲リスト
- * @param onRangeSelect 範囲選択時のコールバック
- * @param modifier Modifier
- */
-@Composable
-fun RangeList(
-    onRangeSelect: (Int) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Column(modifier = modifier) {
-        Text(
-            text = stringResource(
-                id = R.string.input_keyword_select_range_explanation
-            ),
-            textAlign = TextAlign.Center,
-            fontSize = 18.sp,
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(8.dp),
-        )
-        RangeListContent(onRangeSelect)
-    }
-}
