+++ feature-keyword/build.gradle.kts
@@ -1,0 +1,74 @@
+plugins {
+    alias(libs.plugins.android.library)
+    alias(libs.plugins.jetbrains.kotlin.android)
+    alias(libs.plugins.ksp)
+    alias(libs.plugins.compose.compiler)
+}
+
+android {
+    namespace = "com.example.feature_keyword"
+    compileSdk = libs.versions.compileSdk.get().toInt()
+
+    defaultConfig {
+        minSdk = libs.versions.minSdk.get().toInt()
+
+        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles("consumer-rules.pro")
+    }
+
+    buildTypes {
+        release {
+            isMinifyEnabled = false
+            proguardFiles(
+                getDefaultProguardFile("proguard-android-optimize.txt"),
+                "proguard-rules.pro"
+            )
+        }
+    }
+    buildFeatures {
+        compose = true
+    }
+    compileOptions {
+        sourceCompatibility = JavaVersion.toVersion(libs.versions.java.get())
+        targetCompatibility = JavaVersion.toVersion(libs.versions.java.get())
+    }
+    kotlinOptions {
+        jvmTarget = libs.versions.kotlin.jvm.get()
+    }
+}
+
+dependencies {
+    implementation(project(":core"))
+    implementation(project(":shared-ui"))
+
+    implementation(libs.androidx.appcompat)
+    implementation(libs.material)
+    implementation(libs.androidx.core.ktx)
+    implementation(libs.androidx.lifecycle.runtime.ktx)
+    implementation(libs.androidx.activity.compose)
+    implementation(platform(libs.androidx.compose.bom))
+    implementation(libs.androidx.ui)
+    implementation(libs.androidx.ui.tooling.preview)
+    implementation(libs.androidx.material3)
+    implementation(libs.androidx.compose.material.iconsExtended)
+    implementation(libs.kotlinx.collections.immutable)
+
+    implementation(libs.androidx.hilt.navigation.compose)
+
+    // Dagger-Hilt
+    implementation(libs.dagger.hilt.android)
+    implementation(libs.androidx.lifecycle.runtime.compose.android)
+    ksp(libs.dagger.hilt.android.compiler)
+
+    // DataStore
+    implementation(libs.androidx.datastore.preferences)
+
+    // MockitoJUnitRunner
+    testImplementation(libs.mockito)
+    testImplementation(libs.junit)
+    testImplementation(libs.dagger.hilt.android.testing)
+    kspTest(libs.dagger.hilt.android.compiler)
+    testImplementation(libs.androidx.runner)
+    testImplementation(libs.androidx.core.testing)
+    testImplementation(libs.kotlinx.coroutines.test)
+}
