+++ feature-keyword/src/main/java/com/example/feature_keyword/mock/MockKeyword.kt
@@ -1,0 +1,8 @@
+package com.example.feature_keyword.mock
+
+import kotlinx.collections.immutable.toImmutableList
+
+object MockKeyword {
+    val sampleHistoryList = listOf("item1", "item2", "item3").toImmutableList()
+    const val KEYWORD = "焼肉"
+}
